feat: Add CLI
