+++ src/Scheduler/AIscheduler.py
@@ -1,11 +1,11 @@
 import datetime
-from typing import List, Dict, Optional, Tuple
-from src.model import Task, TimeSlot
-from src.Scheduler.scheduler import CalendarManager, SlotScore, SlotScorer
+from typing import List, Dict, Optional
+from src.models import Task, TimeSlot
+from src.scheduler.CalendarManager import CalendarManager
+from src.scheduler.SlotScorer import SlotScorer
 
-# --- AI Agent ---
 
-class TodoListAIScheduler:
+class AIScheduler:
     def __init__(self, settings: Optional[Dict] = None):
         if settings is None:
             settings = {
@@ -41,7 +41,7 @@ class TodoListAIScheduler:
         tasks_already_scheduled_today = [task for task in self.tasks if task.scheduled_start and task.scheduled_start.date() == target_date.date()]
 
         # 2. Sắp xếp các task chờ xử lý theo độ ưu tiên và deadline
-        pending_tasks.sort(key=lambda t: (t.priority.value, t.due_date if t.due_date else datetime.datetime.max), reverse=True)
+        pending_tasks.sort(key=lambda t: (-t.priority.value, t.due_date if t.due_date else datetime.datetime.max))
 
         # 3. Cập nhật thông tin dự án cho SlotScorer
         self.slot_scorer.update_scheduled_tasks_for_projects(tasks_already_scheduled_today)
@@ -82,9 +82,6 @@ class TodoListAIScheduler:
             else:
                 print(f"Không tìm thấy slot phù hợp cho task: {task.description}")
         
-        # Cập nhật lại danh sách tasks chính
-        self.tasks = [t for t in self.tasks if t.scheduled_start is not None] + pending_tasks # Kết hợp task đã có và task mới lên lịch
-
     def get_schedule_for_date(self, date: datetime.datetime) -> List[Task]:
         return sorted(
             [task for task in self.tasks if task.scheduled_start and task.scheduled_start.date() == date.date()],
